requirements.txt
flask
flask_cors
flask_login
paho-mqtt
slackclient
